Fix reply marker line
