+++ app/src/main/java/dev/baseio/discordjetpackcompose/ui/routes/dashboard/main/chatscreen/ChatMessageItem.kt
@@ -1,19 +1,18 @@
 package dev.baseio.discordjetpackcompose.ui.routes.dashboard.main.chatscreen
 
 import android.annotation.SuppressLint
-import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.text.InlineTextContent
-import androidx.compose.foundation.text.appendInlineContent
 import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.ListItem
 import androidx.compose.material.ModalBottomSheetState
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -21,19 +20,9 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Path
-import androidx.compose.ui.graphics.StrokeJoin
-import androidx.compose.ui.graphics.StrokeJoin.Companion
-import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.Placeholder
-import androidx.compose.ui.text.PlaceholderVerticalAlign
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import coil.compose.rememberAsyncImagePainter
 import coil.request.ImageRequest
 import coil.transform.RoundedCornersTransformation
@@ -59,7 +48,7 @@ fun ChatMessageItem(
         replyToMessage = message.replyToMessage
       )
     }
-    ListItem(
+    ChatListItem(
       modifier = Modifier
         .padding(0.dp)
         .combinedClickable(
@@ -71,75 +60,36 @@ fun ChatMessageItem(
             }
           }
         ),
-      icon = {
-        ImageBox(
-          Modifier.size(40.dp),
-          imageUrl = "https://images.unsplash.com/photo-1464863979621-258859e62245?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=686&q=80"
-        )
-      },
-      text = {
-        ChatTitle(message)
-      },
-      secondaryText = {
-        ChatBody(message)
-      }
+      imageSize = 55.dp,
+      message = message
     )
   }
 }
 
-@OptIn(ExperimentalMaterialApi::class)
 @Composable
-fun MessageItemReplyHeader(
-  replyToMessage: String
+fun ChatListItem(
+  modifier: Modifier,
+  imageSize: Dp,
+  message: DiscordMessageEntity
 ) {
-  ListItem(
-    modifier = Modifier
-      .padding(0.dp),
-    icon = {
-      ReplyLine()
-    },
-    text = {
-      ChatReplyTitle(
-        toUserName = "Person",
-        message = replyToMessage
-      )
-    }
-  )
-}
-
-@Composable
-fun ReplyLine() {
-  Canvas(
-    modifier = Modifier
-      .size(50.dp)
-      .padding(0.dp)
+  Row(
+    modifier = modifier.fillMaxWidth(),
+    verticalAlignment = Alignment.CenterVertically
   ) {
-    val width = size.width
-    val height = size.height
-    val path = Path().apply {
-      moveTo(
-        x = width.times(1.2f),
-        y = height.times(.45f)
-      )
-      cubicTo(
-        x1 = width.times(0f),
-        y1 = height.times(.4f),
-
-        x2 = width.times(.35f),
-        y2 = height.times(1.5f),
-
-        x3 = width.times(.30f),
-        y3 = height.times(1.4f)
-      )
-    }
-    drawPath(
-      path = path,
-      color = Color.White.copy(alpha = .90f),
-      style = Stroke(
-        width = 2f,
-        join = StrokeJoin.Miter
-      )
+    ImageBox(
+      Modifier
+        .size(imageSize)
+        .padding(8.dp),
+      imageUrl = "https://images.unsplash.com/photo-1464863979621-258859e62245?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=686&q=80"
     )
+    Column(
+      horizontalAlignment = Alignment.Start,
+      modifier = Modifier.padding(start = 8.dp)
+    ) {
+      ChatTitle(message)
+      Spacer(modifier = Modifier.height(4.dp))
+      ChatBody(message)
+    }
   }
 }
 
@@ -181,58 +131,14 @@ fun ChatTitle(message: DiscordMessageEntity) {
       style = MessageTypography.h1.copy(
         color = DiscordColorProvider.colors.textPrimary
       ),
-      modifier = Modifier.padding(end = 8.dp)
+      modifier = Modifier.padding(top = 0.dp, bottom = 0.dp, end = 8.dp)
     )
     Text(
       message.createdDate.calendar().formattedFullDateTime(),
       style = MessageTypography.overline.copy(
         color = DiscordColorProvider.colors.textSecondary.copy(alpha = 0.8f)
-      )
-    )
-  }
-}
-
-@Composable
-fun ChatReplyTitle(
-  toUserName: String,
-  message: String
-) {
-  val annotatedString = buildAnnotatedString {
-    appendInlineContent(id = "imageId")
-    pushStyle(
-      SpanStyle(
-        fontWeight = FontWeight.Medium,
-        color = DiscordColorProvider.colors.textPrimary
-      )
-    )
-    append("   $toUserName   ")
-    pushStyle(
-      SpanStyle(
-        fontWeight = FontWeight.Normal,
-        color = DiscordColorProvider.colors.textSecondary.copy(alpha = 0.8f)
-      )
-    )
-    append(message)
-  }
-  val inlineContentMap = mapOf(
-    "imageId" to InlineTextContent(
-      Placeholder(20.sp, 20.sp, PlaceholderVerticalAlign.TextCenter)
-    ) {
-      ImageBox(
-        modifier = Modifier.size(32.dp),
-        imageUrl = "https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=687&q=80"
-      )
-    }
-  )
-  Row(verticalAlignment = Alignment.CenterVertically) {
-    Text(
-      text = annotatedString,
-      inlineContent = inlineContentMap,
-      style = MessageTypography.h1.copy(
-        color = DiscordColorProvider.colors.textPrimary,
-        fontSize = 14.sp
       ),
-      modifier = Modifier.padding(end = 8.dp)
+      modifier = Modifier.padding(top = 0.dp, bottom = 0.dp)
     )
   }
 }
